T-Doll-related function refactor
